Sauvegarde tampon pour redémarrage du docker VSCODE

+++ Forker.pm
@@ -17,8 +17,7 @@ my %options = (
     destroy   => 'yes',
 );
 
-my $self;
-my %self;               # Références vers l'instance singleton de la classe.
+my $self;               # Références vers l'instance singleton de la classe.
 my $mainPID;            # ID du process servant de process pères aux différents Forks fils.
 my @forkChilds = ();    # Liste permettant le ressencement de tous les forks fils.
 my $buffer;             # Référence vers une table de hash partagée.
@@ -41,6 +40,7 @@ sub new { my ($class, $buffer) = @_;
     #    return $self;
     #}
     bless ($this, $class);
+    $this->{buffer} = $buffer || {};
     return $this;
 }
 
@@ -59,24 +59,42 @@ sub new { my ($class, $buffer) = @_;
 
 =cut
 sub exec { my ($this, $function, @args) = @_;
-    if not defined $mainPID {
+    if (not defined $mainPID) {
         $mainPID = fork() or die($!);
-        startProcess($function, @args) if not $mainPID;
+        _startProcess($function, @args) if not $mainPID;
     } else {
-        addFork($function, @args);
+        _addFork($function, @args);
     }
 }
 
 sub DESTROY {
     # Tuer le processus père et tous les processus fils
-    kill -$mainPID;
+    kill -$mainPID if (defined $mainPID);
 }
 
 # Méthode utils privée
-my $startProcess = sub { my ($function, @args) = @_;
+sub _startProcess { my ($function, @args) = @_;
+    # Vérifie si le premier paramètre est bien une référence sur fonction
+    die "ERREUR -> le premier paramètre de la fonction doit être une reference sur fonction !" if (ref $function != 'CODE');
+
+    # Démarrage du traitement dans un nouveau fork
+    my $mainPID = fork or die($!);
+    if (not $mainPID) {
+        # Child process
+        $function->(@args);
+    } else {
+        # Ajout du traitement
+        _addFork($function, @args);
+    }
+}
+
+sub _addFork { my ($function, @args) = @_;
+    # Vérifie si le premier paramètre est bien une référence sur fonction
+    die "ERREUR -> le premier paramètre de la fonction doit être une reference sur fonction !" if (ref $function != 'CODE');
+
     # Démarrage du traitement dans un nouveau fork
     my $PID = fork or die($!);
-    if not $PID {
+    if (not $PID) {
         # Child process
         $function->(@args);
     } else {
